+++ test/algorithm-prime-modulus-ring.cpp
@@ -1,13 +1,19 @@
-#include <rain/algorithm/algorithm.hpp>
-#include <rain/algorithm/mod-ring.hpp>
+#include <rain/algorithm/fibonacci.hpp>
+#include <rain/algorithm/prime-modulus-ring.hpp>
 
 #include <cassert>
 #include <iostream>
 
 int main() {
 	{
-		Rain::Algorithm::ModRing<long long, 998244353> x, y(100);
-		Rain::Algorithm::ModRing<long long, 1000000009> z;
+		using PMR1 = Rain::Algorithm::PrimeModulusRing<
+			long long,
+			Rain::Algorithm::PrimeModulus<long long, 998244353>>;
+		using PMR2 = Rain::Algorithm::PrimeModulusRing<
+			long long,
+			Rain::Algorithm::PrimeModulus<long long, 1000000009>>;
+		PMR1 x, y(100);
+		PMR2 z;
 		assert(sizeof(long long) >= 8);
 		x += 5;
 		assert(x == 5);
@@ -45,11 +51,24 @@ int main() {
 		y = 2LL * 3 * 5 * 7 * 11 * 13 * 17 * 23 * 29 * 31;
 		assert(x / y / 10000 / 3049857272LL / 31 == 39486758);
 		assert(x / 10000 / 3049857272LL / 39486758 / 31 == y);
+
+		// Factorials and chooses.
+		PMR1::precomputeFactorials(4096);
+		assert(PMR1::factorials()[5] == 120);
+		assert(PMR1::factorials()[1000] == 421678599);
+		assert(PMR1(6).choose(2) == 15);
+		assert(PMR1(1000).choose(45) == 991398900);
 	}
 
 	{
-		Rain::Algorithm::ModRing<std::size_t, 998244353> x, y(100);
-		Rain::Algorithm::ModRing<std::size_t, 1000000009> z;
+		Rain::Algorithm::PrimeModulusRing<
+			std::size_t,
+			Rain::Algorithm::PrimeModulus<std::size_t, 998244353>>
+			x, y(100);
+		Rain::Algorithm::PrimeModulusRing<
+			std::size_t,
+			Rain::Algorithm::PrimeModulus<std::size_t, 1000000009>>
+			z;
 		// Make sure this is true!
 		assert(sizeof(std::size_t) >= 8);
 		x += 5;
@@ -90,7 +109,10 @@ int main() {
 		assert(x / 10000 / 3049857272LL / 39486758 / 31 == y);
 	}
 
-	Rain::Algorithm::ModRing<uint64_t, 998244353> v{99};
+	Rain::Algorithm::PrimeModulusRing<
+		uint64_t,
+		Rain::Algorithm::PrimeModulus<uint64_t, 998244353>>
+		v{99};
 	assert(sizeof(uint64_t) == 8);
 	assert((v + 504957) / 3 == 168352);
 	assert(
@@ -100,13 +122,17 @@ int main() {
 	assert(v - 998244355 == 97);
 	std::cout << "v is " << v << '.' << std::endl;
 
-	auto res = Rain::Algorithm::ModRing<std::size_t, 23>::power(5, 4);
+	auto res = Rain::Algorithm::PrimeModulusRing<
+							 std::size_t,
+							 Rain::Algorithm::PrimeModulus<std::size_t, 23>>(5)
+							 .power(4);
 	assert(res == 625 % 23);
 
 	// Fibonacci tests.
 	assert(
-		(Rain::Algorithm::fibonacciNumber<
-			 Rain::Algorithm::ModRing<std::size_t, 988244353>>(1000000000) ==
+		(Rain::Algorithm::fibonacciNumber<Rain::Algorithm::PrimeModulusRing<
+			 std::size_t,
+			 Rain::Algorithm::PrimeModulus<std::size_t, 988244353>>>(1000000000) ==
 		 910643820));
 
 	return 0;