+++ test/networking-client.cpp
@@ -20,14 +20,15 @@ int main() {
 			StreamTypeInterface,
 			TcpProtocolInterface,
 			NoLingerSocketOption>
-			client("facebook.com:80");
+			client("google.com:80");
 		std::cout << "Connected to " << client.peerHost() << std::endl;
+		// HTTP 0.9.
 		client.send("GET /\r\n");
 
 		std::string buffer(1_zu << 10, '\0');
 		client.recv(buffer);
 		std::cout << buffer << std::endl;
-		assert(buffer.substr(0, 30) == "HTTP/1.1 301 Moved Permanently");
+		assert(buffer.substr(0, 15) == "HTTP/1.0 200 OK");
 
 		auto timeElapsed = std::chrono::steady_clock::now() - timeBegin;
 		std::cout << "Time elapsed: " << timeElapsed << "." << std::endl